feat(cli): Add 'helm deployment delete'
Closes #239
